+++ src/main/kotlin/ucb/accounting/backend/mapper/CustomerMapper.kt
@@ -9,7 +9,7 @@ class CustomerMapper {
         fun entityToDto(customer: Customer): CustomerDto {
             return CustomerDto(
                 customerId = customer.customerId,
-                subaccountId = customer.subaccountId,
+                subaccountId = customer.subaccountId.toLong(),
                 prefix = customer.prefix,
                 displayName = customer.displayName,
                 firstName = customer.firstName,
